+++ src/main/java/com/techelevator/snacks/Gum.java
@@ -1,8 +1,10 @@
 package com.techelevator.snacks;
 
+import java.math.BigDecimal;
+
 public class Gum extends Snack {
 
-    public Gum (String snackName, double snackPrice) {
+    public Gum (String snackName, BigDecimal snackPrice) {
         super(snackName, snackPrice);
 
     }
