+++ src/test/java/com/techelevator/snacksTests/DrinkTest.java
@@ -5,12 +5,14 @@ import com.techelevator.snacks.Drink;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 public class DrinkTest {
 
     @Test
     public void get_snack_message_drink_returns_glug_glug_yum() {
         //arrange
-        Drink sevenUp = new Drink("", 0.0);
+        Drink sevenUp = new Drink("", BigDecimal.valueOf(0.00));
 
         //act
         String actualMessage = sevenUp.getSnackMessage();
