+++ src/test/java/com/techelevator/snacksTests/GumTest.java
@@ -5,12 +5,14 @@ import com.techelevator.snacks.Gum;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 public class GumTest {
 
     @Test
     public void get_snack_message_gum_returns_chew_chew_yum() {
         //arrange
-        Gum bazooka = new Gum("", 0.0);
+        Gum bazooka = new Gum("", BigDecimal.valueOf(0.0));
 
         //act
         String actualMessage = bazooka.getSnackMessage();
